requirements.txt
python-dotenv==1.0.1
firecrawl_py==1.12.0
pydantic==1.10.14
streamlit==1.42.2
openai==1.65.1
pandas==2.2.3